modift compile errors

+++ ft_memset.c
@@ -6,7 +6,7 @@
 /*   By: hyyang <hyyang@student.42seoul.kr>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/01/24 00:45:53 by hyyang            #+#    #+#             */
-/*   Updated: 2021/01/24 02:21:18 by hyyang           ###   ########.fr       */
+/*   Updated: 2021/01/24 03:13:52 by hyyang           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,7 +15,7 @@
 void	*ft_memset(void *b, int c, size_t len)
 {
 	unsigned char	*p;
-	int				i;
+	size_t			i;
 
 	p = b;
 	i = 0;
